+++ src/utils/__tests__/styleSearch-test.js
@@ -125,6 +125,26 @@ test("ignores matches within double-quote strings", t => {
   /* eslint-enable quotes */
 })
 
+test("`checkStrings` option", t => {
+  t.deepEqual(styleSearchResults({
+    source: "abc 'abc'",
+    target: "b",
+    checkStrings: true,
+  }), [ 1, 6 ])
+  t.end()
+})
+
+test("`withinStrings` option", t => {
+  /* eslint-disable quotes */
+  t.deepEqual(styleSearchResults({
+    source: 'abc "abc"',
+    target: "b",
+    withinStrings: true,
+  }), [6])
+  /* eslint-enable quotes */
+  t.end()
+})
+
 test("ignores matches within comments", t => {
   t.deepEqual(styleSearchResults({
     source: "abc/*comment*/",
@@ -134,6 +154,24 @@ test("ignores matches within comments", t => {
     source: "abc/*command*/",
     target: "a",
   }), [0])
+  t.end()
+})
+
+test("`checkComments` option", t => {
+  t.deepEqual(styleSearchResults({
+    source: "abc/*abc*/",
+    target: "b",
+    checkComments: true,
+  }), [ 1, 6 ])
+  t.end()
+})
+
+test("`withinComments` option", t => {
+  t.deepEqual(styleSearchResults({
+    source: "abc/*abc*/",
+    target: "b",
+    withinComments: true,
+  }), [6])
   t.end()
 })