+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Head
 
+- Added: `string-no-newline` rule.
 - Added: `unspecified: "bottomAlphabetical"` option for `rule-properties-order`.
 - Added: `stylelint-disable-line` feature.
+- Added: `withinComments`, `withinStrings`, and `checkStrings` options to `styleSearch`, and `insideString` property to the `styleSearch` match object.
 - Fixed: `font-weight-notation` does not throw false warnings when `normal` is used in certain ways.
 
 # 4.4.0